==> bioinformatics/3_load_transformCounts.Rmd <==
---
title: "Normalization and Transformation of scRNA-Seq Counts"
output:
  md_document:
    variant: markdown_github
    toc: true
  html_document:
    toc: true
    toc_float: true
---

```{r prereqs}
knitr::opts_chunk$set(echo = T, message = FALSE, warning = FALSE)
stdt <- date()

# Packages
library(tidyverse)
library(Seurat)
library(kableExtra)

# Set data directory path
datadir <- "/hpc/group/chsi-mic-2022/data/Christian2021CellReports/"


# Set output directory path (be sure that this directory exists and that
# to have write permission to it).
procdir <- "/work/owzar001/MIC2022/PROC/"
```

**Goal of this workshop:** Learn how to import multiple cellranger output data sets into R, perform QC and transform UMI counts.

**What's covered in this workshop:**

-   Import four cellranger count matrices R as Seurat objects
-   Peform basic QC on the objects
-   Transform UMI counts

**Data source:** Christian et al. 2021 Cell Reports (<https://doi.org/10.1016/j.celrep.2021.109118>)

Note that the Seurat - Guided Clustering Tutorial (<https://satijalab.org/seurat/articles/pbmc3k_tutorial.html>) includes substantial overlapping information.

**METHOD DETAILS, Single-cell RNA-seq analysis, page e4**

-   Pre-processing: "Raw short reads were demultiplexed, filtering and mapped to mouse genome GRCm38/mm10 using cellranger v2.02."

-   QC: "The gene count matrices from cellranger were subjected to quality control, pre-processing and clustering using the R Seurat 2.3.4 package (Butler et al., 2018). Low-quality cells that had less than 200 expressed genes and more than 5% mitochondrial genes were filtered out."

-   Analysis: "Gene counts were scaled to total gene expression and percentage of mitochondrial genes with a scaling factor of 10,000, and then log-transformed. The high dimensional data for each sample were reduced by PCA and t-Distributed Stochastics Neighbor Embedding (tSNE). We used the FindCluster function to group clusters in each sample with a resolution of 0.6. Differential expressed genes (DEGs) were identified using the Wilcoxon rank-sum test."

## 1. Load data into R as Seurat objects

First, we identify the hdf5 (\*.h5) files containing the UMI counts from cellranger from the Christian et al. 2021 paper, and create a manifest including file and sample ids and the md5 hash. Review the R list.files(), basename() and gsub() functions before reviewing this code chunk.

```{r datafiles}

# Get all files with ,h5 extension
h5files <- list.files(path = datadir, pattern = ".h5", full.names = TRUE)
# Inspect inventory
tibble::tibble(
  fname = basename(h5files),
  sampid = gsub("_10x.h5", "", fname),
  md5 = tools::md5sum(h5files)
) -> manifest
manifest
```

Confirm that you have identified data files corresponding (see Fig 2A of Christian et al. for tSNE plots of each sample type):

-   Tem = effector memory in tumor
-   Tcm = central memory in tumor
-   Trm = resident memory in tumor
-   disTrm = resident memory T-cells outside of the tumor

Next, we write a function to import the hdf5 files as Seurat objects. Review the Read10X_h5() and CreateSeuratObject() functions from the Seurat package and the file.path() and paste0() functions from R before reviewing this code chunk.

```{r importfun}
readin_10x_to_SO <- function(datadir, fname, sampid, min.cells = 0, min.genes = 0) {
  # Create a full file name
  filename <- file.path(datadir, paste0(sampid, "_10x.h5"))
  Read10X_h5(filename) %>%
    CreateSeuratObject(
      project = sampid,
      min.cells = min.cells,
      min.genes = min.genes
    ) -> seu.obj
  return(seu.obj)
}
```

Next, we use our custom function to read in the hdf5 files. This step will create a list of four Seurat objects by looping over the rows of the manifest data frame. Review list and dataframe R objects and how to extract information from these, and the R seq_len() function before reviewing this code chunk. Note that the number of genes and cells vary across the files. It should be noted that one can use R apply type function instead of a loop (we will demonstrate the application of R lapply function later).

```{r importh5}
seulist <- list()
for (i in seq_len(nrow(manifest))) {
  fname <- manifest[[i, "fname"]]
  sampid <- manifest[[i, "sampid"]]
  seulist[[i]] <- readin_10x_to_SO(datadir, fname, sampid)
  names(seulist)[i] <- sampid
}
seulist
```

The following illustrates how to access a specific object from the list you have created

```{r listaccessex}
seulist[["Tem"]]
```

## 2. Review the imported data against information provided in the Christian paper

*METHOD DETAILS, Single-cell RNA-seq analysis, page e4*: "Low-quality cells that had less than 200 expressed genes and more than 5% mitochondrial genes were filtered out."

First, we observe that the features are gene symbols by inspecting one of the objects

### Adding mitochondrial content percentages to the meta data

Use the R rowname() function to extract the feature names.

```{r features}
tem <- seulist$Tem
rownames(tem)[1:10]
```

The mitochondrial genes can be identified as follows. Review the R grep() function before reviewing this code chunk.

```{r mtfeatures}
rownames(tem)[grep("^mt-", rownames(tem))]
```

Next, we add the mitochondrial content percentage to each Seurat objects in our list (see QC notebook for details). We consider to equivalent approaches (looping and lapply) for updating Seurat objects in a list.

First, we define a function which augments the meta data of a Seurat object by the mitochondrial proportion.

```{r addmtfun}
addmt <- function(so) {
  so[["percent.mt"]] <- PercentageFeatureSet(so, pattern = "^mt-")
  return(so)
}
```

Next we apply this function by looping over the list.

```{r addmtloop}
seulist.mt <- list()
for (i in 1:length(seulist)) {
  seulist.mt[[i]] <- addmt(seulist[[i]])
}
names(seulist.mt) <- names(seulist)
seulist.mt
```

Generally speaking the preferred approach to applying functions to a list of Seurat objects is to use the R lapply() function.

```{r addmtapply}
seulist <- lapply(seulist, addmt)
seulist
```

### Generate QC plots

Next, we examine the distribution of nFeature and nCount of each object

```{r plotfun1}
plotfun1 <- function(so) {
  so@meta.data %>%
    ggplot(aes(x = nCount_RNA, y = nFeature_RNA, color = percent.mt)) +
    geom_point(size = 1) +
    labs(x = "nCount", y = "nFeature", color = "MT%") +
    theme_classic() +
    geom_hline(yintercept = 200, linetype = 2) +
    ggtitle(so@project.name)
}
```

This function can be used to produce the plot for a single object.

```{r plotfun1ex}
plotfun1(seulist[["Tem"]])
```

The function can also be applied to all Seurat objects in the list.

```{r plotfun1apply}
lapply(seulist, plotfun1)
```

Are there any cells with less than 200 expressed genes?

We first write a function to produce a figure

```{r plotfun2}
plotfun2 <- function(so) {
  so@meta.data %>%
    ggplot(aes(x = percent.mt)) +
    geom_histogram() +
    theme_classic() +
    geom_vline(xintercept = 5, linetype = 2) +
    ggtitle(so@project.name)
}
```

```{r}
lapply(seulist, plotfun2)
```

Are there any cells with more than 5% mitochondrial genes?

### Tabulating QC summaries

One can use the summarize() function from the dplyr function to produce summary statistics for the meta data. The following is a simple demonstration.

```{r summ1}
seulist[["Tem"]]@meta.data %>%
  summarize(
    n = n(),
    min = min(nCount_RNA),
    mean = mean(nCount_RNA),
    median = median(nCount_RNA),
    max = max(nCount_RNA)
  ) %>%
  mutate(dlabel = "Tem", .before = "n")
```

To generalize, we can consider writing a function

```{r summ2}
summfun1 <- function(so) {
  so@meta.data %>%
    summarize(
      n = n(),
      min = min(nCount_RNA),
      mean = mean(nCount_RNA),
      median = median(nCount_RNA),
      max = max(nCount_RNA)
    ) %>%
    mutate(dlabel = so@project.name, .before = "n")
}

summfun1(seulist[["Tem"]])
```

To generalize this further, so as to not hardcode the meta data column, we can consider the following function. Note the use of the !!sym() function to programmatically pass the variable name.

```{r summ3}
summfun2 <- function(so, mdcol) {
  so@meta.data %>%
    summarize(
      n = n(),
      min = min(!!(sym(mdcol))),
      mean = mean(!!sym(mdcol)),
      median = median(!!sym(mdcol)),
      max = max(!!sym(mdcol))
    ) %>%
    mutate(dlabel = so@project.name, mfeat = mdcol, .before = "n")
}
```

Now apply this function.

```{r summ4}
summfun2(seulist[["Tem"]], "nCount_RNA")
summfun2(seulist[["Tem"]], "nFeature_RNA")
summfun2(seulist[["Tem"]], "percent.mt")
```

The do

```{r summ5}
do.call("rbind", lapply(seulist, summfun2, mdcol = "percent.mt"))
```

Summarize as a table

```{r}
df <- data.frame(seuobj = names(seulist))
df[, c("min", "mean", "max")] <- NA
for (i in 1:length(seulist)) {
  df[i, c("min", "mean", "max")] <- seulist[[i]]@meta.data %>%
    summarize(
      min = min(nFeature_RNA),
      mean = mean(nFeature_RNA),
      max = max(nFeature_RNA)
    )
}
df
```

An alternative way to do this is by using purrr mapping functions

```{r}
seulist %>%
  map_dfr(~ .x@meta.data, .id = "seuobj") %>%
  group_by(seuobj) %>%
  summarize(
    min = min(nFeature_RNA),
    mean = mean(nFeature_RNA),
    max = max(nFeature_RNA)
  )
```

## 3. Transform gene counts

*METHOD DETAILS, Single-cell RNA-seq analysis, page e4*: "Gene counts were scaled to total gene expression and percentage of mitochondrial genes with a scaling factor of 10,000, and then log-transformed."

### Examine slots of a Seurat object

We will examine the structure of a Seurat object. First, we get a list of the slots

```{r slotnames}
tem <- seulist$Tem
slotNames(tem)
```

The object has a single assay

```{r assay}
tem@assays
```

It is also the "active" assay

```{r activeassay}
tem@active.assay
```

Now let's look at the slots of the RNA assay

```{r}
slotNames(tem@assays$RNA)
```

This will extract the count matrix

```{r}
tem@assays$RNA@counts[1:5, 1:5]
```

Note that the data slot is identical to the counts slot

```{r}
identical(tem@assays$RNA@data, tem@assays$RNA@counts)
```

### Normalize the counts

The Seurat NormalizedData() function can be used to "normalize" the UMI counts.

```{r}
tem.s <- NormalizeData(
  object = tem,
  normalization.method = "LogNormalize",
  scale.factor = 10000
)
```

Note that while the count slot remains unchanged,

```{r}
identical(tem.s@assays$RNA@counts, tem@assays$RNA@counts)
tem.s@assays$RNA@counts[1:5, 1:5]
```

the data object has been modified

```{r}
identical(tem.s@assays$RNA@data, tem@assays$RNA@data)
tem.s@assays$RNA@data[1:10, 1:10]
```

The following function can be used to normalize all objects in the list.

```{r normalizefun}
normfun <- function(so) {
  NormalizeData(
    so,
    normalization.method = "LogNormalize",
    scale.factor = 10000
  )
}
```

Apply function to the list.

```{r normalizedata}
seulist <- lapply(seulist, normfun)
seulist
```

Another widely used transformation is provided by the Seurat SCTransform() (regularized negative binomial regression to normalize UMI count data) function.

## 4. Save the list of normalized Seurat objects

```{r saveobj}
newfile <- file.path(procdir, "seulist-n.rds")
saveRDS(seulist, file = newfile)
tools::md5sum(newfile)
```
```{r sessioninfo}
sessionInfo()
```
